Move alignment conveniences to a separate file
